% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{getMetaData}
\alias{getMetaData}
\title{Get a list of possible meta data tags}
\usage{
getMetaData(dev = FALSE)
}
\arguments{
\item{dev}{should the development webservce be called}
}
\description{
List all meta data tags available to annotate a repository or data object
}
