% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAFtokens.R
\name{getToken}
\alias{getToken}
\title{get user token via query login}
\usage{
getToken(username)
}
\arguments{
\item{username}{the ices login name of the user}
}
\value{
a named list contianing the j web token and the expiration date
}
\description{
get user token via query login
}
\examples{
\dontrun{
jwt <- getToken("colin")
jwt
}
}
