% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesTAFDB-package.R
\docType{package}
\name{icesTAFDB-package}
\alias{icesTAFDB-package}
\alias{icesTAFDB}
\title{Transparent Assessment Framework (TAF) Database Web Services}
\description{
R interface to access the web services of the ICES Transparent Assessment
Framework (TAF) Database <http://taf.ices.dk>.
}
\references{
ICES Transparent Assessment Framework: \url{http://taf.ices.dk}.
}
\author{
Colin Millar.
}
