% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesTAF-package.R
\docType{package}
\name{icesTAFWeb-package}
\alias{icesTAFWeb-package}
\alias{icesTAFWeb}
\title{Transparent Assessment Framework (TAF) Web Services}
\description{
R interface to access the web services of the ICES Transparent Assessment
Framework (TAF) web application <http://taf.ices.dk>.
}
\references{
ICES Transparent Assessment Framework: \url{http://taf.ices.dk}.
}
\author{
Colin Millar.
}
