% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.update.R
\name{taf.update}
\alias{taf.update}
\title{Update a TAF stock assessment}
\usage{
taf.update(message, jwt, path = ".")
}
\arguments{
\item{message}{a commit message for the changes being submitted}

\item{jwt}{a valid token identifying the user}

\item{path}{the path to the local repository (default = ".")}
}
\value{
a named list contianing the j web token and the expiration date
}
\description{
Update a TAF stock assessment
}
\examples{
\dontrun{
jwt <- getToken("colin")
jwt
}
}
