% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf_assessments.R
\name{taf_assessments}
\alias{taf_assessments}
\title{get a list of public taf assessments}
\usage{
taf_assessments(year, stock, type, what = NULL, jwt = NULL)
}
\arguments{
\item{year}{the year the assessments were conducted}

\item{stock}{the ices stock code}

\item{type}{the type of analysis, assessment, forecast,
exploratory, etc.}

\item{what}{should the function return results, scripts, or a summary,
Default is summary.}

\item{jwt}{a token aquired by running getToken}
}
\value{
a named list
}
\description{
get a list of public taf assessments
}
\examples{
\dontrun{
assessments <- taf_assessments()
jwt
}
}
