% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf_projects.R
\name{taf_projects}
\alias{taf_projects}
\title{List all available TAF projects}
\usage{
taf_projects()
}
\description{
List all TAF projects and related information that is available
to the requesting user.  All requests are sent with a token so
that the web service can identify which projects are visible.
}
