% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Show user information}
\usage{
test(use_token = TRUE, verbose = FALSE, quiet = FALSE, dev = FALSE)
}
\arguments{
\item{use_token}{should a user token be checked}

\item{verbose}{should verbose output be shown}

\item{quiet}{minimal output to user or none}

\item{dev}{should the development webservce be called}
}
\description{
Return a list of information for the logged in user
}
